Renamed to shi7_learning.
