+++ shi7/shi7.py
@@ -60,7 +60,6 @@ def make_arg_parser():
     parser.add_argument('-trim_q', '--trim_qual', help='Trim read ends until they reach trim_q (default: %(default)s)', default=32, type=int)
     parser.add_argument('--drop_r2', help='When combining FASTAs, drop R2 reads and remove "R1" from read name (helpful with non-stitchable PE reads) (default: False)', choices=[True, False], default='False', type=convert_t_or_f)
     parser.set_defaults(shell=False, single_end=False)
-
     return parser
